refactor common template utility into a package
