+++ dan_layer/engine/tests/common/Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 [package]
-name = "state"
+name = "common"
 version = "0.1.0"
 edition = "2021"
 
@@ -8,7 +8,6 @@ edition = "2021"
 
 [dependencies]
 tari_template_abi = { path = "../../../template_abi" }
-common = { path = "../common" }
 
 [profile.release]
 opt-level = 's'     # Optimize for size.
